Corrected reports that use SQL locate() with ContainerPath
"locate(path2.path, path1.path) > 0" is incorrect, so was replaced with
"path1.path like path2.path || '/%'" in several reports. For example,
say we have a path2.path equal to "6", the first condition will say that
if path2.path equals "126/32" then it matches. This is not what we want.
We want to find any paths that match regexp "^6/", not regexp "6", so only
true children of the given path are matched.