Eager Load Regular Reports
If a postProcess method exists for an AbstractReport derived class
and the method accesses associations of an object that have not been
loaded, then an Exception might be thrown.

This is because the "nextChunk" of an AbstractBiobankListProxy is
loaded in a new thread, with a new session, so the 1001st Object and
on will reference a Session that is closed and throw an exception
when a lazily loaded association or list is accessed.