Many modifications to make transactions supported by ModelWrappers
* add getIdProperty() method to ModelWrapper and update BaseWrapperBuilder
to automatically generate the correct code using the corresponding Peer
class. Use this property to set and get the wrapped object's id instead
of doing it with reflection.
* add getPersistTasks() and getDeleteTasks() methods to be overriden
by ModelWrappers to supply the TaskList of what to do in order to persist
and delete the wrapped objects, correspondingly
* be careful about when to fire property changes (don't do it if the
property has not been initialized [since the old value is not necessarily
correct] and don't fire a property change if the new and old value are
in fact the same).
* make ModelWrapper.notifyListeners() package accessible so it can be
called after a Delete or Persist action is successfully committed.
* add initialization of property check methods to ModelWrapper
* add client-side length checks just before a commit()
* add internal class to Property since TypeReference cannot always be
serialized. The internal class pulls out all necessary info when it is
instantiated and is Serializable. Also add "isCollection()" method to
the Property class. This makes the Property class Serializable.
* implement transaction-ability for StudyWrapper
* keep a StudyWrapper's Study's StudyEventAttr collection up-to-date
* don't do HQL queries on new objects (specifically StudyEventAttr) as
hibernate doesn't like that and throws an exception. Check if isNew()
before sending HQL.
* include internal package for WrapperKlazz lookup (in ModelWrapperHelper)
* @Override WritableApplicationServiceImpl.prepareRequest() so
BiobankSessionAction-s do not have their request replaced with something
else
* clean up the look of BiobankORMDAOImpl as query() method was wayyyyyy
too big
* Update Property generation files to make Accessor implementation
Serializable
* Add a Cascade class to make cascading properties simple :-)
* Switch from *ExampleQuery-s to BiobankSessionAction-s so that the
session can be directly accessed and new object's returned as part
of the returned graph.
* A TON of other work to neatly abstract checks and persist/ delete
actions.