Tons of Changes to Make Wrappers Transaction-able
DONE (1) enforce bidirectional links (affect old only if property initialized)
DONE (2) specimen <-> collectionEvent two links - careful
DONE (3) assume only addTo and removeFrom are called, then remember oldValues
and persist the removedValues (if instructed to do so in getPersistTasks)
DONE (4) whether a collection is loaded or not, the wrapper should remove old
values in the collection
SKIPPED, IN FAVOR OF LAZY APPLICATION (5) run through UML and remove links that have HUGE collections.
NUTS
	* center.getProcessingEventCollection
	* center.specimenCollection
	* center.originInfoCollection
KINDA NUTS
	* center.srcDispatchCollection
	* center.dstDispatchCollection
	* center.requestCollection
	* study.patientCollection
SORTA NUTS
	* site.containerCollection
FINE
	* container.specimenPositionCollection
- REPLACE these with HQL queries, probably cached?
DONE (6) remember elements removed from or added to a collection and persist them
if the holder of the collection is persisted. Why? Because when would you ever
not want to save the collection "properly"? Why add or remove elements, then
save the holder if you don't want them to be saved?
DONE (7) if added to collection, persist; if removed, require programmer to define
whether to delete or persist the element.
DONE (8) DO NOT persistAdded, persistRemoved, or deletedRemoved automatically
TODO (10) Do we need to set inverse on the getters? Or is this done by CaCORE/ hibernate?
For example, does ce.getSpecimenCollection().get(0).getCollectionEvent().getSpecimenCollection()
load the collectionEvent again, or the specimen collection again?
DONE (11) Maybe don't set bidirectional links if the inverse is not loaded.