Use DateTimeWidget Validator on Dummy Observable Value
When backspace is pressed, the createdAt property is attempted to be
set to null, but the validator doesn't allow this. When the form is
reset, the createdAt property is "changed" to the same value it used
to be. Because the value has not changed, the appearance was not updated.

We think the createdAt value goes from 1970 -> null -> 1970, but it
really just stays at 1970 because the validator won't allow it to be
null.

resolves #1167