added log-debug property to ant build file
Before this change, the 'ant -Ddebug=true' command would enable the server to
listen for connections on port 8080 and enable log4j loggers on the server side.
This causes a lot of debugging information to be generated in the jboss server
logs and causes the application to run slower. Running the test suite takes a
lot longer.

Now a different property is used to enable the log4j loggers: log-debug. To
enable the loggers now use the following ant command:

 - ant -Dlog-debug=true

To enable port 8080 and enable log4j debugging on server use this command:

 - ant -Ddebug=true -Dlog-debug=true