Make Getters Bi-Directional
* made getters set the inverse direction (if the property the getter was
called on WAS NOT CACHED). So the inverse is set only once.
* getters should be bidirectional, specifically for lazily-loaded associations.
Otherwise, problems can occur, as seen in the following example: assume two
objects A and B have a bi-directional link. If A is loaded, then A.getB() is
called, then B.getA() is called, the session cache should have been used to
find and use the correct A object. HOWEVER, if before B.getA() was called,
someone changed B's value of A, then B.getA() would return a different id
object than the database, and A.getB().getA() would not return the original A.

resolves #1448