Use isInitialized Instead of isPropertyCached on Internal Setters
For example, if a new ProcessingEvent is created and a new Specimen
as well, then the Specimen.setProcessingEvent() is called with the
new ProcessingEvent, then the ProcessingEvent's specimenCollection
will be null because it is new. If it's null then it is initialized,
so we should be able to immediately add the Specimen to it (because
it is new). If we used isPropertyCached, then the
ProcessingEventWrapper's specimenCollection property would not be
cached and we would not immediately add to it. This is a particular
problem when accessing the model underneath directly. For example, if
isPropertyCached() returns false, but the model's property is
accessed (via property.get()) then a null value can be returned
instead of accessing the database.