Stop using ApplicationService.search(), see comments
Switch to query(DetachedCriteria) instead because search() was using
an object (e.g. new Container()) to contain values to filter for. This
is a problem because now Container has primitive values defined and
by default activityStatus = ActivityStatus.ACTIVE; so, when a new
one was created, search() would look for id = x and activityStatus =
ACTIVE.

So, DetachedCriteria is now used instead to search for only the
added Criterion (id only in ModelWrapper.reload() -> getObjectFromDatabase()).