Two bug fixes related to Certificates.
1. Changed location for saving cacerts file.

The following exception was happening on MS Windows 7 when the client
was installed in C:\Program Files (x86):

java.io.FileNotFoundException:
  C:\Program Files (x86)\BioBank\cacerts (Access is denied)

The location for saving the cacerts file was changed to the user's home
directory and the file renamed to "biobank_cacerts".

2. Fix for null pointer exception

Variable CustomTrustManager.chain was not initialized. Therefore, when
CustomTrustManager.getRejectedIssuers() was called, it would generate
an exception.